+++ example/src/App.tsx
@@ -1,6 +1,5 @@
-import * as React from 'react';
-
 import { FoldingFeatureProvider } from '@logicwind/react-native-fold-detection';
+
 import SampleScreen from './SampleScreen';
 
 export default function App() {