+++ ios/FoldDetection.mm
@@ -1,19 +1,24 @@
 #import "FoldDetection.h"
 
 @implementation FoldDetection
-RCT_EXPORT_MODULE()
 
-// Example method
-// See // https://reactnative.dev/docs/native-modules-ios
-RCT_EXPORT_METHOD(multiply:(double)a
-                  b:(double)b
-                  resolve:(RCTPromiseResolveBlock)resolve
-                  reject:(RCTPromiseRejectBlock)reject)
+- (void)startListening
 {
-    NSNumber *result = @(a * b);
-
-    resolve(result);
 }
 
+- (void)stopListening
+{
+}
+
+- (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:
+    (const facebook::react::ObjCTurboModule::InitParams &)params
+{
+    return std::make_shared<facebook::react::NativeFoldingFeatureSpecJSI>(params);
+}
+
++ (NSString *)moduleName
+{
+    return @"FoldingFeature";
+}
 
 @end