+++ src/context/FoldingFeatureContext.tsx
@@ -1,11 +1,5 @@
 import type { PropsWithChildren } from 'react';
-import React, {
-  createContext,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-} from 'react';
+import { createContext, useContext, useEffect, useMemo, useState } from 'react';
 import { Platform } from 'react-native';
 
 import { subscribeToFoldingFeature } from './subscribeToFoldingFeature';